+++ CHANGES.rst
@@ -14,6 +14,7 @@ development
 - [doc] Improve installation documentation
 - [mw] Direct access to OPS 3.1 API has been disabled, so stop linking to it
 - [ui] Fix: When creating new project, convert project name to string
+- [mw] Improve DpmaRegisterAccess re. language selection
 
 2018-03-17 0.150.0
 ==================